requirements.txt
django
django-cors-headers
djangorestframework
psycopg2-binary
dj-database-url
gunicorn
